pyproject.toml
[tool.poetry]
name = "email-to-whatsapp-mom-bills"
version = "0.1.0"
description = ""
authors = ["Vinicius999 <viniciussilva.dsc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
imap-tools = "^1.7.0"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
pywhatkit = "^5.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
